Combined all createListing parts into one.  Currently not functional -> needs image uploading and testing for current parts
