+++ src/components/createListing/createListing.css
@@ -121,6 +121,152 @@
   margin: auto;
   animation: spin 0.8s infinite linear;
 }
+
+/* Location */
+.listing-wrapper {
+  display: flex;
+}
+
+.listing-inputs {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+}
+
+.label-text {
+  margin-top: 4px;
+  padding-right: 25px;
+  font-weight: bold;
+  font-size: 120%;
+  color: #4d4842;
+}
+
+.gen-subsec {
+  display: flex;
+  padding-top: 8px;
+}
+.inputBox {
+  margin-bottom: 3%;
+  padding-left: 2%;
+  height: 30px;
+  width: 100%;
+  border-radius: 15px;
+  border: 2px solid rgb(210 210 210 / 80%);
+  background-color: white;
+  color: #4d4842;
+  font-size: 12px;
+}
+
+.inputBox:hover {
+  background-color: #00b1821d;
+}
+
+.stateInputBox {
+  vertical-align: middle;
+  margin-bottom: 3%;
+}
+
+
+/* Details */
+.details-wrapper {
+  display: inline-block;
+  width: 200px;
+  align-content: center;
+}
+
+.overall-details {
+  margin-bottom: 8%;
+}
+
+.beds {
+  display: flex;
+}
+
+.baths {
+  display: flex;
+}
+
+.maxppl {
+  display: flex;
+}
+
+.input-label-details {
+  display: flex;
+  align-items: center;
+  font-weight: bold;
+  font-size: 120% !important;
+  padding: 5%;
+  color: #4d4842;
+}
+
+.input-box-details {
+  width: 100%;
+  padding-left: 6%;
+  background-color: white;
+  border-radius: 5px;
+  height: 35px;
+  border-radius: 20px;
+  margin-top: 3%;
+  margin-bottom: 3%;
+  border: 2px solid rgb(210 210 210 / 47%);
+  color: #4d4842;
+}
+
+.input-box-details:hover {
+  background-color: #00b1821d;
+}
+
+
+/* Amenities */
+.amenities-checkbox {
+  height: auto;
+}
+
+.amenities-list {
+  display: flex;
+  justify-content: space-between;
+}
+
+.no-amenities-check {
+  float: left !important;
+  margin-top: 5px !important;
+}
+
+.amenity {
+  margin-right: 10px;
+}
+
+/* Photos */
+.single-img-container {
+  position: relative;
+}
+
+.single-img-container .delete-btn {
+  position: absolute;
+  transform: translate(-100%, 0%);
+  -ms-transform: translate(-50%, -50%);
+  background-color: #ea5d5d;
+  color: white;
+  font-size: 16px;
+  width: 20px;
+  padding: 12px 20px 10px 10px;
+  border: none;
+  cursor: pointer;
+  border-radius: 5px;
+}
+
+.single-img-container .delete-btn:hover {
+  background-color: red;
+}
+
+input[type="file"] { display: none; }
+
+.upload-file {
+  color: #00b183;
+  border: solid 2px #00b183;
+  padding: 6px;
+}
+
 @keyframes spin {
   from {
     transform: rotate(0deg);
