--- src/components/createListing/detailsListing.css
@@ -1,47 +1,0 @@
-.details-wrapper {
-  display: inline-block;
-  width: 200px;
-  align-content: center;
-}
-
-.overall-details {
-  margin-bottom: 8%;
-}
-
-.beds {
-  display: flex;
-}
-
-.baths {
-  display: flex;
-}
-
-.maxppl {
-  display: flex;
-}
-
-.input-label-details {
-  display: flex;
-  align-items: center;
-  font-weight: bold;
-  font-size: 120% !important;
-  padding: 5%;
-  color: #4d4842;
-}
-
-.input-box-details {
-  width: 100%;
-  padding-left: 6%;
-  background-color: white;
-  border-radius: 5px;
-  height: 35px;
-  border-radius: 20px;
-  margin-top: 3%;
-  margin-bottom: 3%;
-  border: 2px solid rgb(210 210 210 / 47%);
-  color: #4d4842;
-}
-
-.input-box-details:hover {
-  background-color: #00b1821d;
-}
