--- src/components/createListing/locationListing.css
@@ -1,42 +1,0 @@
-.listing-wrapper {
-  display: flex;
-}
-
-.listing-inputs {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-}
-
-.label-text {
-  margin-top: 4px;
-  padding-right: 25px;
-  font-weight: bold;
-  font-size: 120%;
-  color: #4d4842;
-}
-
-.gen-subsec {
-  display: flex;
-  padding-top: 8px;
-}
-.inputBox {
-  margin-bottom: 3%;
-  padding-left: 2%;
-  height: 30px;
-  width: 100%;
-  border-radius: 15px;
-  border: 2px solid rgb(210 210 210 / 80%);
-  background-color: white;
-  color: #4d4842;
-  font-size: 12px;
-}
-
-.inputBox:hover {
-  background-color: #00b1821d;
-}
-
-.stateInputBox {
-  vertical-align: middle;
-  margin-bottom: 3%;
-}
