--- src/components/createListing/photos/photos.css
@@ -1,21 +1,0 @@
-.single-img-container {
-  position: relative;
-}
-
-.single-img-container .delete-btn {
-  position: absolute;
-  transform: translate(-100%, 0%);
-  -ms-transform: translate(-50%, -50%);
-  background-color: #ea5d5d;
-  color: white;
-  font-size: 16px;
-  width: 20px;
-  padding: 12px 20px 10px 10px;
-  border: none;
-  cursor: pointer;
-  border-radius: 5px;
-}
-
-.single-img-container .delete-btn:hover {
-  background-color: red;
-}
